% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Who_is_ZZ.R
\name{Who_is_ZZ}
\alias{Who_is_ZZ}
\title{Zhou Zhou}
\usage{
Who_is_ZZ(printlog = TRUE)
}
\arguments{
\item{printlog}{logical. Whether to print.}
}
\description{
Who is Zhou Zhou?
}
\examples{
Who_is_ZZ()
}
\keyword{Zhou}
